Renamed Sequence back to Collection.
